requirements.txt
beautifulsoup4==4.12.2
cfscrape==2.1.1
httpx==0.24.1
nonebot2==2.0.0
Pillow==9.5.0
playwright==1.35.0
urllib3==1.26.16
requests==2.28.2
